+++ HW5/app/src/main/java/com/kodeco/android/countryinfo/ui/components/CountryInfoRow.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -15,13 +14,11 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
 import com.kodeco.android.countryinfo.types.Country
-import com.kodeco.android.countryinfo.ui.viewModels.CountryInfoViewModel
+import com.kodeco.android.countryinfo.ui.screens.countryList.CountryInfoViewModel
 
 @Composable
 fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
@@ -38,8 +35,7 @@ fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
                     if (isExpanded) {
                         // add to close
                         viewModel.incrementCloses()
-                    }
-                    else {
+                    } else {
                         // add to open
                         viewModel.incrementOpens()
                     }
@@ -67,7 +63,7 @@ fun CountryInfoRow(viewModel: CountryInfoViewModel, country: Country) {
 
 
             if (isExpanded) {
-                AsyncImage(model = country.flags.png, contentDescription = "Flag", contentScale = ContentScale.FillWidth, modifier = Modifier.fillMaxWidth())
+                CountryDetailsScreen(country = country)
             }
 
         }
