+++ HW5/app/src/main/java/com/kodeco/android/countryinfo/ui/screens/countryList/CountryInfoViewModel.kt
@@ -1,26 +1,15 @@
-package com.kodeco.android.countryinfo.ui.viewModels
+package com.kodeco.android.countryinfo.ui.screens.countryList
 
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.ViewModelProvider
 import com.kodeco.android.countryinfo.types.Country
 import com.kodeco.android.countryinfo.types.ICountriesRepository
-import com.kodeco.android.countryinfo.ui.components.Resource
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-
-class CountryInfoViewModelFactory (private val repository: ICountriesRepository) : ViewModelProvider.NewInstanceFactory() {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return CountryInfoViewModel(repository) as T
-    }
-}
 
 
 class CountryInfoViewModel(private var repository: ICountriesRepository): ViewModel() {
@@ -32,10 +21,6 @@ class CountryInfoViewModel(private var repository: ICountriesRepository): ViewMo
                 _timerFlow.value += 1
             }
         }
-
-//        viewModelScope.launch {
-//            reloadData()
-//        }
     }
 
     private val _countries = MutableStateFlow(emptyList<Country>())
@@ -59,8 +44,6 @@ class CountryInfoViewModel(private var repository: ICountriesRepository): ViewMo
     private val _closes = MutableStateFlow<Int>(0)
     val closes = _closes.asStateFlow()
 
-
-
     fun reloadData() {
         viewModelScope.launch {
             _reloadData()
@@ -80,16 +63,17 @@ class CountryInfoViewModel(private var repository: ICountriesRepository): ViewMo
         _countries.value = emptyList()
         _reloadCount.value += 1
 
-//        viewModelScope.launch {
         repository.getAllCountries().collect {
             when (it) {
                 is Resource.Error -> {
                     _errorMessage.value = it.message
                     _isLoading.value = false
                 }
+
                 is Resource.Loading -> {
                     _isLoading.value = true
                 }
+
                 is Resource.Success -> {
                     _errorMessage.value = null
                     _isLoading.value = false
@@ -97,9 +81,5 @@ class CountryInfoViewModel(private var repository: ICountriesRepository): ViewMo
                 }
             }
         }
-
-//        }
-
     }
-
 }
